schema 3.0: restructuring de-serializer for compatibility
