+++ .gitignore
@@ -11,3 +11,4 @@ builtins
 
 node_modules
 package-lock.json
+/.editor_settings
