--- colyseus/.editorconfig
@@ -1,6 +1,0 @@
-[*.{lua}]
-end_of_line = lf
-charset = utf-8
-trim_trailing_whitespace = true
-indent_style = space
-indent_size = 2
