+++ example/testsuite.script
@@ -9,9 +9,9 @@ local http = require "test.http"
 function init(self)
 
   deftest.add(schema_serializer)
-  deftest.add(client)
-  deftest.add(auth)
-  deftest.add(http)
+  -- deftest.add(client)
+  -- deftest.add(auth)
+  -- deftest.add(http)
 
   -- deftest.add(fossil_delta)
   -- deftest.add(delta_container)
