+++ game.project
@@ -1,14 +1,11 @@
 [project]
 title = Colyseus Defold Client
-version = 0.15
-dependencies#0 = https://github.com/britzl/deftest/archive/master.zip
-dependencies#1 = https://github.com/defold/extension-websocket/archive/master.zip
+version = 0.16
+dependencies#0 = https://github.com/britzl/deftest/archive/refs/tags/2.8.0.zip
+dependencies#1 = https://github.com/defold/extension-websocket/archive/refs/tags/4.0.0.zip
 
 [bootstrap]
 main_collection = /example/testsuite.collectionc
-
-[input]
-use_accelerometer = 0
 
 [display]
 width = 1280
@@ -19,10 +16,4 @@ shared_state = 1
 
 [library]
 include_dirs = colyseus
-
-[physics]
-scale = 0.02
-
-[facebook]
-appid = 353169041992501
