+++ server/package.json
@@ -18,8 +18,8 @@
   },
   "homepage": "https://github.com/colyseus/create-colyseus#readme",
   "devDependencies": {
-    "@colyseus/loadtest": "^0.15.2",
-    "@colyseus/testing": "^0.15.2",
+    "@colyseus/loadtest": "^0.16.0-preview.0",
+    "@colyseus/testing": "^0.16.0-preview.0",
     "@types/cors": "^2.8.6",
     "@types/express": "^4.17.1",
     "@types/mocha": "^8.2.3",
@@ -30,13 +30,13 @@
     "typescript": "^5.0.4"
   },
   "dependencies": {
-    "@colyseus/core": "^0.15.0-preview.8",
-    "@colyseus/monitor": "^0.15.0-preview.3",
-    "@colyseus/playground": "^0.15.10",
-    "@colyseus/redis-driver": "^0.15.0-preview.0",
-    "@colyseus/redis-presence": "^0.15.0-preview.2",
-    "@colyseus/tools": "^0.15.0-preview.10",
-    "colyseus": "^0.15.0-preview.4",
+    "@colyseus/core": "^0.16.0-preview.0",
+    "@colyseus/monitor": "^0.16.0-preview.0",
+    "@colyseus/playground": "^0.16.0-preview.0",
+    "@colyseus/redis-driver": "^0.16.0-preview.0",
+    "@colyseus/redis-presence": "^0.16.0-preview.0",
+    "@colyseus/tools": "^0.16.0-preview.0",
+    "colyseus": "^0.16.0-preview.0",
     "cors": "^2.8.5",
     "express": "^4.16.4"
   }
