+++ test/schema/InheritedTypes/Entity.lua
@@ -2,7 +2,7 @@
 -- THIS FILE HAS BEEN GENERATED AUTOMATICALLY
 -- DO NOT CHANGE IT MANUALLY UNLESS YOU KNOW WHAT YOU'RE DOING
 -- 
--- GENERATED USING @colyseus/schema 1.0.0-alpha.58
+-- GENERATED USING @colyseus/schema 3.0.0-alpha.42
 -- 
 
 local schema = require 'colyseus.serialization.schema.schema'
